integrates tableau.py into network_summary.py and removed tableau.py
